requirements.txt
streamlit==1.38.0
requests==2.32.3
pydantic==2.9.2
python-dotenv==1.0.1

# RAG / LLM tooling (all OSS-friendly)
langchain==0.3.7
langchain-community==0.3.7
faiss-cpu==1.8.0.post1
sentence-transformers==3.2.1
pypdf==5.1.0